requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
crontab==1.0.1
cryptography==41.0.0
Deprecated==1.2.14
Django==4.2.1
django-debug-toolbar==4.1.0
django-dump-die==0.1.5
django-oauth-toolkit==2.2.0
django-otp==1.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.42.0
geographiclib==2.0
googlemaps==4.10.0
gunicorn==21.2.0
idna==3.4
jwcrypto==1.5.0
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.7.0
pyotp==2.8.0
pypng==0.20220715.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
wrapt==1.15.0